requirements.txt
aiogram>=3.20.0
yt-dlp>=2024.12.1
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
jinja2>=3.1.4
aiofiles>=24.1.0
sqlalchemy[asyncio]>=2.0.36
aiosqlite>=0.20.0
python-dotenv>=1.0.1
pydantic>=2.10.0
pydantic-settings>=2.6.0
itsdangerous>=2.2.0
humanize>=4.11.0
apscheduler>=3.10.4
